Fix a bug with output dtype in Fock backend and support sorting FockState lexicographically (#40)
* fix a bug with per_norm dtype

* remain double after sqrt

* typo

+++ src/deepquantum/photonic/circuit.py
@@ -312,7 +312,7 @@ class QumodeCircuit(Operation):
             if self._state_expand is not None:
                 unitary = torch.block_diag(unitary, torch.eye(1, dtype=unitary.dtype, device=unitary.device))
             sub_mats = vmap(sub_matrix, in_dims=(None, None, 0))(unitary, state, final_states)
-            per_norms = self._get_permanent_norms(state, final_states)
+            per_norms = self._get_permanent_norms(state, final_states).to(unitary.dtype)
             if is_prob:
                 rst = vmap(self._get_prob_fock_vmap)(sub_mats, per_norms)
             else:
@@ -730,13 +730,13 @@ class QumodeCircuit(Operation):
         if state.ndim == 1:
             sub_mat = sub_matrix(unitary, state, final_state)
             per = permanent(sub_mat)
-            amp = per / self._get_permanent_norms(state, final_state)
+            amp = per / self._get_permanent_norms(state, final_state).to(per.dtype)
         else:
             idx_nonzero = torch.where(torch.sum(state, dim=-1) == torch.sum(final_state))[0]
             amp = torch.zeros(state.shape[0], dtype=unitary.dtype, device=unitary.device)
             if idx_nonzero.numel() != 0:
                 sub_mats = vmap(sub_matrix, in_dims=(None, 0, None))(unitary, state[idx_nonzero], final_state)
-                per_norms = self._get_permanent_norms(state[idx_nonzero], final_state)
+                per_norms = self._get_permanent_norms(state[idx_nonzero], final_state).to(unitary.dtype)
                 rst = vmap(self._get_amplitude_fock_vmap)(sub_mats, per_norms).flatten()
                 amp[idx_nonzero] = rst
         return amp
@@ -1028,7 +1028,7 @@ class QumodeCircuit(Operation):
         """
         final_states = self._all_fock_basis
         sub_mats = vmap(sub_matrix, in_dims=(None, None, 0))(unitary, init_state, final_states)
-        per_norms = self._get_permanent_norms(init_state, final_states)
+        per_norms = self._get_permanent_norms(init_state, final_states).to(unitary.dtype)
         rst = vmap(self._get_prob_fock_vmap)(sub_mats, per_norms)
         state_dict = {}
         prob_dict = defaultdict(list)
